Move tests to integration test dir
