+++ extendr-api/tests/graphics_tests.rs
@@ -1,6 +1,10 @@
+use std::fmt::Write;
+
 use extendr_api::graphics::color::predefined::{antiquewhite, black, darkkhaki, deepskyblue};
 use extendr_api::graphics::color::Color;
-use extendr_api::graphics::{Context, Device, FontFace, Unit};
+use extendr_api::graphics::{
+    Context, DevDesc, Device, DeviceDescriptor, DeviceDriver, FontFace, R_GE_gcontext, Raster, Unit,
+};
 use extendr_api::prelude::*;
 
 #[test]
@@ -90,5 +94,180 @@ fn graphics_test() {
             std::thread::sleep(std::time::Duration::from_millis(2000));
             assert!(false); // did we check in the flag as "false?"
         }
+    }
+}
+
+// Taking the mutable references so that we can peek at the values from
+// outside, which is (only?) useful for testing.
+struct TestDevice<'a> {
+    last_mode: &'a mut i32,
+    value: &'a mut f64,
+    canvas: &'a mut String,
+    closed: &'a mut bool,
+}
+
+impl<'a> DeviceDriver for TestDevice<'a> {
+    fn activate(&mut self, _: DevDesc) {
+        *self.value = 100.0;
+    }
+
+    fn mode(&mut self, mode: i32, _: DevDesc) {
+        *self.last_mode = mode;
+        *self.value += 1.0;
+    }
+
+    fn close(&mut self, _: DevDesc) {
+        *self.closed = true;
+    }
+
+    fn new_page(&mut self, _: R_GE_gcontext, _: DevDesc) {
+        self.canvas.clear();
+    }
+
+    fn clip(&mut self, from: (f64, f64), to: (f64, f64), _dd: DevDesc) {
+        let (f_x, f_y) = from;
+        let (t_x, t_y) = to;
+        writeln!(
+            *self.canvas,
+            "clip from=({f_x:.1}, {f_y:.1}), to=({t_x:.1}, {t_y:.1})"
+        )
+        .unwrap();
+    }
+
+    fn circle(&mut self, center: (f64, f64), r: f64, _: R_GE_gcontext, _: DevDesc) {
+        let (x, y) = center;
+        writeln!(*self.canvas, "circle center=({x:.1}, {y:.1}), r={r:.1}").unwrap();
+    }
+
+    fn line(&mut self, from: (f64, f64), to: (f64, f64), _: R_GE_gcontext, _: DevDesc) {
+        let (f_x, f_y) = from;
+        let (t_x, t_y) = to;
+        writeln!(
+            *self.canvas,
+            "line from=({f_x:.1}, {f_y:.1}), to=({t_x:.1}, {t_y:.1})"
+        )
+        .unwrap();
+    }
+
+    fn rect(&mut self, from: (f64, f64), to: (f64, f64), _: R_GE_gcontext, _: DevDesc) {
+        let (f_x, f_y) = from;
+        let (t_x, t_y) = to;
+        writeln!(
+            *self.canvas,
+            "rect from=({f_x:.1}, {f_y:.1}), to=({t_x:.1}, {t_y:.1})"
+        )
+        .unwrap();
+    }
+
+    fn text(
+        &mut self,
+        pos: (f64, f64),
+        str: &str,
+        rot: f64,
+        hadj: f64,
+        _: R_GE_gcontext,
+        _: DevDesc,
+    ) {
+        let (x, y) = pos;
+        writeln!(
+            *self.canvas,
+            "text pos=({x:.1}, {y:.1}) str='{str}' rot={rot:.1} hadj={hadj:.1}"
+        )
+        .unwrap();
+    }
+
+    fn raster<T: AsRef<[u32]>>(
+        &mut self,
+        raster: Raster<T>,
+        pos: (f64, f64),
+        size: (f64, f64),
+        rot: f64,
+        interpolate: bool,
+        _: R_GE_gcontext,
+        _: DevDesc,
+    ) {
+        let (x, y) = pos;
+        let (width, height) = size;
+        let Raster { pixels, width: w } = raster;
+        let pixels_str = pixels
+            .as_ref()
+            .iter()
+            .map(|&p| format!("{p}"))
+            .collect::<Vec<String>>()
+            .join("|");
+
+        writeln!(
+                *self.canvas,
+                "raster {pixels_str} w={w} pos=({x:.1}, {y:.1}) size=({width:.1}, {height:.1}) rot={rot:.1} interpolate={interpolate}"
+            )
+            .unwrap();
+    }
+}
+
+#[test]
+fn graphic_device_test() {
+    test! {
+        let mut value = 0.0;
+        let mut last_mode = 0;
+        let mut closed = false;
+        let mut canvas = String::new();
+
+        let device_driver = TestDevice {
+            last_mode: &mut last_mode,
+            value: &mut value,
+            canvas: &mut canvas,
+            closed: &mut closed,
+        };
+
+        let device_descriptor = DeviceDescriptor::new();
+        let device = device_driver.create_device::<TestDevice>(device_descriptor, "test device");
+
+        let gc = Context::from_device(&device, Unit::Device);
+
+        // if activate() is invoked, value should be 100.0
+        assert_eq!(value, 100.0);
+        assert!(!closed);
+
+        device.mode_on().unwrap();
+
+        // if mode() is invoked, value and last_mode should be updated
+        assert_eq!(last_mode, 1);
+        assert_eq!(value, 101.0);
+
+        device.mode_off().unwrap();
+
+        // if mode() is invoked, value and last_mode should be updated
+        assert_eq!(last_mode, 0);
+        assert_eq!(value, 102.0);
+
+        device.clip((1.1, 2.2), (3.3, 4.4), &gc);
+        device.circle((1.1, 2.2), 3.3, &gc);
+        device.line((1.1, 2.2), (3.3, 4.4), &gc);
+        device.rect((1.1, 2.2), (3.3, 4.4), &gc);
+        // x element of `center` is `hadj`, a horizontal adjustment
+        device.text((1.1, 2.2), "foo", (0.5, 0.0), 5.5, &gc);
+
+        let r = Raster {
+            pixels: &[1, 2, 3, 4, 5, 6],
+            width: 3,
+        };
+        device.raster(r, (1.1, 2.2), (3.3, 4.4), 5.5, false, &gc);
+
+        assert_eq!(canvas, "clip from=(1.1, 2.2), to=(3.3, 4.4)\n\
+                                circle center=(1.1, 2.2), r=3.3\n\
+                                line from=(1.1, 2.2), to=(3.3, 4.4)\n\
+                                rect from=(1.1, 2.2), to=(3.3, 4.4)\n\
+                                text pos=(1.1, 2.2) str='foo' rot=5.5 hadj=0.5\n\
+                                raster 1|2|3|4|5|6 w=3 pos=(1.1, 2.2) size=(3.3, 4.4) rot=5.5 interpolate=false\n\
+                                ");
+
+        // Clearing canvas.
+        device.new_page(&gc);
+        assert_eq!(canvas, "");
+
+        // check if the R doesn't crash on closing the device.
+        R!("dev.off()")?;
+
+        assert!(closed);
     }
 }
